# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Aurora UI API",
    "version": "1.0.0",
    "description": "Machine-readable access to Aurora UI copy-paste React components for Next.js. Fetch the registry first, then component metadata or source. See docs/FOR_AI.md for integration steps.",
    "contact": {
      "url": "https://github.com/Rydaguy101/aurora-ui"
    }
  },
  "servers": [
    {
      "url": "https://aurora-ui-tau.vercel.app",
      "description": "Production docs site"
    }
  ],
  "paths": {
...
    "/api/search": {
      "get": {
        "operationId": "searchComponents",
        "summary": "Search components by keyword or category",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Search slug, title, description, tags"
          },
          {
            "name": "category",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Filter by category name"
          }
        ],
        "responses": {
          "200": {
            "description": "Matching components",
            "content": {
              "application/json": {
                "schema": { "type": "object" }
              }
            }
          }
        }
      }
    },
...
  }
}